apps/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Vaibhav Maurya",email = "vibhu.astute@gmail.com"}
]

requires-python = ">=3.12"
dependencies = [
    "portia-sdk-python>=0.6.2,<0.7",
    "httpx>=0.28.1,<0.29.0",
    "pydantic>=2.11.7,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "wikipedia-api>=0.6.0,<1.0.0"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
